[MOV/DEL] documents_spreadsheet_bundle: move back in documents_spreadsheet
The Split between the 2 modules (documents_spreadsheet, documents_spreadsheet_bundle)
was introduced for clarify the split of lazy loaded files from the other files.

There are 2 issues introduced with this split:

it introduced some weird dependency map between the modules: documents_spreadsheet_bundle
depends on documents_spreadsheet even though the latter cannot work
without the former.
The split is not that clear and is somewhat faulty: some code that can be lazy loaded is
in documents_sreapdsheets - some templates are separated from the compnent definition file
We are now adding features that will depend on Accounting and the dependency mapping will be
even more complicated. All-in-all, it seems like it's not really necessary and we can split
the files in different folders inside a single module.

Task 2809220

closes odoo/enterprise#25707

Related: odoo/upgrade#3386
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>