[FIX] l10n_ar_reports: UnicodeEncodeError error
We receive a traceback error when trying to export a vat book sales and
we have an invoice of a partner that have ” character in the partner
name.

The error shown is because we can not encode the partner name in the TXT
fiiles we need to generate to then be loaded to AFIP:
UnicodeEncodeError(latin-1' codec can't encode character '\u201d'.

With this change we avoid this traceback error by simple skip the
non-encodable characters.

closes odoo/enterprise#26361

X-original-commit: 000822c705dd4c87b983d46a827718721a1fb93f
Signed-off-by: Josse Colpaert <jco@odoo.com>