[FIX] account_budget : fix budget suggestion
To reproduce
============

- create new analytic account
- add a budget

when selecting the budget *Budget 2023: Pessimistic* the line will be set to *Budget 2023: Optimistic*

Purpose
=======

an improvement was added to suggest a budget when dates are set, so when we select *Budget 2023: Pessimistic*
date is set to `from : 01/01/2023 to : 12/31/2023`, so the suggestion function is triggred, and it search for
budgets that have verify a condition on their dates.
in our case *Budget 2023: Pessimistic* and *Budget 2023: Optimistic* have same dates so they will both show up
in the search, then we suggest the first result which is *Budget 2023: Optimistic*.

Specification
=============

to solve the issue we verify if a budget is already set and its dates verify the condition,
then we don't suggest anything.

opw-2802610

closes odoo/enterprise#26347

X-original-commit: b46321e6884f76b1270ed8d63866d51e45c898f9
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>
Signed-off-by: abla001 <abla@odoo.com>