[FIX] l10n_lu_reports_electronic_xml_2_0: fix Partner VAT Intra report
Task: 2823466

Currently, both PSI and LIC declarations are exported in the XML of Partner VAT Intra report even if there is nothing to report. Customers try to upload the XML on LU portal and it is not accepted because one of the declarations is empty.
Should be: If there are only goods or services invoiced, but not both of them, only one of the declarations should be exported accordingly.

closes odoo/enterprise#26254

X-original-commit: b5de7f6d024c68e8c7a36abf95cabcc8c39cf921
Signed-off-by: Quentin De Paoli <qdp@odoo.com>