[FIX] sign: reduces signature dialog opening time
In order to prevent the signature dialog from opening multiple times when
clicking multiple times, a debounce method was added. At the time this was
the best option we could find, but turns out the debounce was making the
dialog take too long to be opened. This commit removes the debounce and uses
a different approach to prevent multiple dialogs from opening. A variable is
used to store if the dialog is opened and when the dialog is destroyed, we set
it to false. This prevents the dialog from opening if there is already a dialog
opened.

Also, as fonts are now cached from 15.3, it was decided to also cache the fonts
in sign requests without sign items.

task-2819572

closes odoo/enterprise#27861

X-original-commit: 5067a960ff9e217ed512833cc4b0d2d9b85c61c3
Signed-off-by: Arnaud Joset <arj@odoo.com>