[REF] account_saft,l10n_*: refactor modules using XSD files loading and validation
Before this commit, all the refactored modules were implementing their own ways of loading XSD files so that they can be used to validate the generated XML files. This was a bit tedious and hard to maintain as all modules almost did the same thing but with different code.

The addition of loading and validation function in tools allowed for a refactoring of every module that used some sort of XSD validation.
This commit refactors all modified modules to unify the loading and validation process of XML/XSD files.

task id=2782053

closes odoo/enterprise#26393

Related: odoo/odoo#89145
Signed-off-by: Laurent Smet <las@odoo.com>