[IMP] web_enterprise: primary assets' order
Following the assumption that enterprise variables were "definitive",
prior to this commit 'primary_variables.scss' and
'bootstrap_overridden.scss' files were injected "on top" of their
community counterparts.

This commit inverts the order and makes use of SCSS '!default' directive
to handle the inheriting process.

The aim is to ease variables overrides across versions and to achieve
that in a more consistent way.

These changes will also pave the way for dynamic variables changes, see
dark-mode.

This commit is part of v16 overall restyle (task-2704984).

Part-of: odoo/enterprise#25737