[FIX] appointment: remove flickering on load when no slot available
When no slots are available or when the appointment is not configured,
the appointment page shows a message with a picure telling that there
are no availabilities for the appointment. Previously, the calendar was
first shown then replaced by the message and the picture, causing a flicker.
The calendar is now no more displayed if no slot is available.

When the appointment is not configured, the alert message is now shown at the
same time as the message telling that there is no slot available.

Task-2812306

closes odoo/enterprise#26837

X-original-commit: c95946bf565e44bd87f2cf71c0ba22bd972ea19c
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>