[FIX] account_batch_payment: fix wizard access rights
Since 15.0, batch payment doesn't require accounting anymore.
Yet the wizard requires the access right that comes with it and is no
longer available with only invoicing like it was in the past.
The setting for it was removed as it was confusing.

Thus, the wizard should only require the Billing role, the same as the
batch payment itself.

opw-2813664

closes odoo/enterprise#26276

X-original-commit: d913457938b9c807da11b0286505b95b58bb99d9
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: Nicolas Viseur <vin@odoo.com>