[FIX] sale_subscription: save payment token in subscription
Before merging subscription with sale orders, when a quotation of a
subscription product was paid in portal, the payment token was already saved to
the subscription automatically. After merging subscription and sale, this
behavior was lost. This commit introduces the code to properly save the payment
token into the subscription if the payment mode is set to 'success_payment'.

task-2824236

X-original-commit: 3e72b96d5d47686f187fd0fffceab1cca01aa0d9
Part-of: odoo/enterprise#27750