[IMP] approvals, sign: introduce view models (step 1)
This commit introduces models that define records being 1:1 map with components,
as a step to move further to having essentially all business code in models.

Having code in models is desirable to have very maintainable code, thanks to
robust and declarative code with an ORM-like architecture.

Task-2831082

closes odoo/enterprise#26637

Related: odoo/odoo#89735
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>