[FIX] sale_subscription: set next_invoice_date = today by default
Before this commit, the next invoice date was set at the end of the first period.
When the first period was invoiced, the next_invoice_date was not updated
for the first invoice. The following invoices would increment the period.

It was not practical and could bring confusions in the code and for the user.
The default next_invoice_date is now equal to "today" and the ui was modified
to make sure the lines were marked as invoiceable. When the user or the cron creates the
first invoice, the period is incremented.

Part-of: odoo/enterprise#27311