[ADD] hr_homeworking: Add employee work location mechanism
PURPOSE
=======

Due to the covid, the home working has been more and more used inside the companies.
Sometimes, due to official recommandations, sometimes by choice of the company.
At the moment, there is no mechanics in Odoo that tell if the employee is working from home.
It can be misleading to refer only to his physical presence to know if he's present.
On the same way, he can be off or in holidays, a physical absence is not equal to homeworking.

task-2700425