[FIX] account_reports: use matching filter logic in front and backend
The enduser can use a search bar in the general ledger report to filter the
accounts. However, the filter logic being used is currently different in the
frontend and the backend.

On the frontend, the account name is split into words, if the start of any of
these words matches the search string, the account is considered a match. In
the backend the word splitting is omitted, so only the start of the complete
name can match the search string. This ambiguous behavior also leads to the
impossibility to match search strings containing spaces in the frontend.

This inconsistency can eventually lead to an error when trying to load the
expanded view of an account that is matched in the frontend, but not in the
backend. When using the search box, the displayed lines will be filtered by the
filter_accounts method in the accountReportsWidget. If the end user now decides
to expand one of the filtered lines, the child lines will be fetched from the
server by calling get_html on the account.general.ledger model including the
search string as a parameter. Because of the inconsistency between the front
and back end logic it is now possible that an account matched in the frontend
will not actually match the search string in the backend. When this happens, a
KeyError occurs when the _do_query explicitly tries to retrieve the child lines
of the expanded account.

Additionally, the filtering logic in the backend will be used when printing or
exporting reports. As such, if the backend yields different results, these will
be reflected in the printed or exported report.

The backend logic was probably the intended logic, while the word splitting in
the frontend was probably only intended to remove the account code (which is
prepended to the display name in the frontend). This can be seen from the
original code included in [commit 1] and modified by [commit 2] where the
backend would do a simple ilike match on any substring in the account name or
code. This generally worked fine, because the results in the backend were
always a superset of those in the frontend. However, even then, the printed
report would not generally reflect the results in the frontend.

Things worsened considerably as of [commit 3], which tried to fix some of the
inconsistencies between frontend and printed or exported reports. The commit
is fundamentally flawed though, restricting the results in the backend far too
much. The exported reports would still not match the frontend, and far more
importantly, expanding any of the results from any search for a prefix of any
but the first word in the account name (you'll probably have to read this a
couple of times...) would now result in the traceback error described above.

The search functionality on the account reports is adapted, so that the result
set matches the one returned by using name_search.

[commit 1]: https://github.com/odoo/enterprise/commit/575e8a4b41d1dcb7189327e723be6dc8379d07ad
[commit 2]: https://github.com/odoo/enterprise/commit/5b000eaa28e8494d2fed043d811d228383bfe704
[commit 3]: https://github.com/odoo/enterprise/commit/a6e0004b46fb2feaa2ae95834dce8a3a8738f0a8

opw-2821275
opw-2821643
opw-2822098
opw-2822231
opw-2822680
opw-2822794
opw-2823255
opw-2823349
opw-2824231
opw-2824287
opw-2824401
opw-2824526
opw-2825405

closes odoo/enterprise#26724

closes odoo/enterprise#27207

X-original-commit: 62c03a2df31187ea87e1931f37b7eda8e0fe251b
Signed-off-by: William André (wan) <wan@odoo.com>