[REF] documents_spreadsheet: don't reload data if never loaded
Add two pivots in two different sheets (but in the same spreadsheet ;))
Add a global filter with a default which applies to both pivots.

The data source for the second pivot is loaded as soon as the spreadsheet is
loaded even though Sheet2 was never displayed.

Task 2830100

Part-of: odoo/enterprise#26290