[FIX] account_avatax{,_sale}: avoid confusion with exempt taxes
It's confusing that taxes that are fully (or partially) exempt still
show on the invoice line as normal. There's no way for the customer to
tell the difference between a (partially) exempt tax and a regular
one. Take the following invoice as an example (tax is excluded from
price in general settings):

|---------+--------------------------+----------|
| Product | Tax                      | Subtotal |
|---------+--------------------------+----------|
| A       | DC STATE TAX [11] (5.0%) | $ 100.00 |
|---------+--------------------------+----------|
|         | Untaxed amount:          | $ 100.00 |
|         | Taxes:                   | $   2.50 |
|         | Total:                   | $ 102.50 |
|---------+--------------------------+----------|

When lines are (partially) exempt the Avatax API returns the normal
tax and it's normal rate but returns a different tax amount. In this
case the normal tax would be $5.00 but because of a partial exemption
the actual tax is $2.50. To avoid confusion this commit hides the tax
names. Curious users can still click on the line to see the tax names
in the popup form view.

This issue does not appear on reports because in both the sale.order
and account.move report the taxes are grouped under the tax group
name.

This is similar to what was done for Taxcloud with
956e2e5cc8c088f70df2109035a55287d9a8e864, although the reasoning for
that was different.

opw-2825425

closes odoo/enterprise#26398

X-original-commit: c6615d4d555fa0a87028e307a62d856031de4af5
Signed-off-by: William André (wan) <wan@odoo.com>