[FIX] account_reports: allow scrollable filter dropdowns
Entries in the filters dropdown lists can overflow when the list is larger than
the screen size allows. Especially the journal and user defined filters are
prone to overflowing. This commit still allows the end user to select any entry
by scrolling the list whenever this is the case.

The css rules introduced in [commit 1] to fix a similar issue with the currency
filter are simply generalized to all account reports filter dropdown lists.

Note: A more general fix might still be warranted. Some minor problems still
remain with this solution. For example, a hard-coded 100px is subtracted to
avoid overlapping the control panel. However, when the list of filter dropdowns
wraps, any scrollable dropdown in the first row will not properly align with the
bottom of the viewport. Additionally,  dropdowns are rendered slightly different
on mobile. Though, at least all filter dropdowns will now be usable by the end
user in all cases.

[commit 1]: https://github.com/odoo/enterprise/commit/4927d8dd350ee1a74d89928115f354c13107729a

opw-2792463

closes odoo/enterprise#27873

X-original-commit: 505fab91d3fbf1e7d0b8424de8b141ceab6b767a
Signed-off-by: De Caluwé Tom (tdc) <tdc@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>