[REF] documents_spreadsheet: fetch metadata only once per model
Concurrent calls to fetch metadata are not "cached".
The promise was cached only once it was resolved.
A concurrent second call would not benefit from the current call.

With this commit, the promise is not "awaited" before being cached.

For the "Pipeline dashboard" template, we go from 8 RPCs to only 2:
4 fields_get and 4 search_read to only one of each.

Task 2830100

closes odoo/enterprise#26290

Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>