[FIX] l10n_mx_edi_report_invoice: remove nested html p tag in p tag for invoicing report
Current behavior:
The invoice payment term of the report is a p tag with another p tag inside like:
<p name="payment_term">
	<span>
		<p> This is a payment term note</p>
	</span>
</p>
but because nested p tags are not supported in html de result in the browser is something like:
<p name="payment_term">
        <span></span>
</p>
<p> This is a payment term note</p>
<p></p>

Similar behavior can be observe for the fiscal position note.

Expected behavior:
The code should respect html format rules and not have nested p tags.

Solution: We replace the parent p tags by div tags, and we modify xml file that have xpath relying
on this tag.

To prevent to breack the xml references we also modified the references from p to div

opw-2804933

closes odoo/enterprise#26372

Related: odoo/odoo#88660
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>