[FIX] documents: Avoid recreating folders/tags/... on update
Purpose
=======

When you upgrade Document app, you get back all data from installation, even if you deleted it.

It can be confusing and create a lot of unwanted noise in your upgrade process.

More over, in the context of the rolling release where you upgrade many times on the year, this can be demotivating.

closes odoo/enterprise#26904

Taskid: 2827001
X-original-commit: 23161f390f00da6b11b0868af55e53e2a618ba6c
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>