[IMP] {website_}appointment: redesign the details view appointment_form
The view has now a column containg a summary of the appointment
(date + tz, length, location) as well as the name, function and
picture (if avatars are activated) of the operator. Name and function
are editable in front end. The name is displayed if the assign_method
is 'chosen' and the user name is not already in the appointment name.

The previous chevrons are clickable. The adress is made multi line.
The view is harmonized with previous ones in the process, as well
as for the footer slightly visible.

Task-2496629

Part-of: odoo/enterprise#24744