[IMP] account_invoice_extract_purchase: apply POs when totals don't match
In the case where we found a Purchase Order but its content doesn't
match the values detected by the OCR, we will now apply the PO without
modification.

Previous behaviour was to use the PO but set all quantities to 0 then
add the results of the OCR at the end. This foten resulted in more
manual work than if we simply set the whole PO without any modification.

closes odoo/enterprise#26719

Task: 2828465
X-original-commit: 0793853c6009b32c882d89cad6b0f4cb9cd7531c
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>