[IMP] l10n_cl_edi_boletas: allow sending a correction for the sales book
Sometimes, we need to be able to resend an already accepted
vatbook e.g. because extra boletas were added.  Maybe, you
forgot to confirm a boleta the day before.

However, we did not have the possibility before to resend
an already accepted vat book.

Also, we removed the constraint that made it impossible to still
send a boleta for a date where the daily sales book was sent.
However, when such a boleta gets accepted, we added a message
in the chatter to indicate that sales book to the user. We also
change the sales book to a new state "to resend".  Then it will
be taken up by the cron and it can also be done from the sales book
itself.

opw-2821628

closes odoo/enterprise#27150

X-original-commit: aff5f3b7c311b3103ca9973fe761617f37697b0f
Signed-off-by: William André (wan) <wan@odoo.com>