[IMP] documents_spreadsheet: add mock server params to  model helper
The test helper `createModelWithDataSource` is improved to allow
custom spreadsheet data, server data and a mockRPC function.

The current easy way of doing this is with `createSpreadsheet[WithPivot| WithList]`
but it instanciates the full spreadsheet UI which is slow and useless
when only testing the model behavior.

Task 2830100

Part-of: odoo/enterprise#26290