[IMP] documents_spreadsheet: rename pivot/list/link
Before this commit, the name of pivots/lists/links were the display name
of the technical model or the action name. When users insert mutilple
pivots/lists/links on the same spreadsheet, it quickly becomes difficulte
to know what pivot/list/link is used for what.

With this commit, users are now able to give a name, either from the
spreadsheet selector dialog or from the side panel.

Task-id 2696631

closes odoo/enterprise#22663

Related: odoo/upgrade#3453
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>