[FIX] account_reports: Bad Debtor status not visible
When going on a Follow-Up reports, you can change the status of a client
(the small ball on the left of their name). In the dropdown, we don't see
the "Bad Debtor" option.

Step to reproduce the issue:
1) Install the Accounting app
2) Go to a Follow-Up reports and try to change their status
You cannot fully see the "Bad Debtor" option.

Solution: During a previous commit (b11141717719744ba143e3d77a5ef883b63c7124),
an `overflow:hidden` was added into the CSS, it should rather be a `overflow:visible`
as the dropdown will overlap with other divs.

opw-2816612

closes odoo/enterprise#26525

X-original-commit: 4b965cc8d45bcf54e567c3d51aeea49f8daa3361
Signed-off-by: Grazioso Andrea (agr) <agr@odoo.com>
Signed-off-by: Desausoi Laurent (lade) <lade@odoo.com>