[IMP] helpdesk: adjust portal visibility followers behavior
Currently, portal followers of a team and its tickets are removed when
the team visibility is changed to something else than portal.
=> We want to keep the behavior, but only when changing the visibility
from portal to something else.

When a follower is added to a ticket in a team in a non-portal visibility,
all of its portal followers are removed.
=> We want to remove this behavior, as it is not intuitive

This commit implements those changes, and also adds a warning when
changing the visibility of the team if that change will automatically
add/remove followers to the team and its tickets.

Related: https://github.com/odoo/odoo/pull/87747

Task-2812551

closes odoo/enterprise#25800

Related: odoo/odoo#87747
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>