[IMP] {website_}appointment: redesign the appointment validation view
Adds a column in the calendar event view accessed from portal or at the
end of the apppointment process (validation / confirmation). This column
consist of a card of name, function and picture of the event organizer,
by default the staff user that has been selected / set on that appointment.
The format is the same as the card in the form view, the user short card
style.

This column only appears if the appointment type has the avatar option
activated and if the user_id is set on the event.

No chevrons are displayed once the validation is reached. The navbar
is still displayed for events having a 'website' category appointment type
linked to them. If an invite is linked as well, the 'all appointments'
button is reworded 'Schedule another appointment', and the redirect url
is used.

Buttons are redesigned as well, and google calendar / outlook icons are
added on them. The footer is slighty visible as in other views (see the
selector o_wappointment_wrap)

As this template can be reached for meetings, one must make sure that
there is an appointment when displaying appointment info in order to
avoid errors. Conditions are added to do so. In particular, the title
is adapted to meeting / appointment and display the event / appointment
name and the user name if not in the event / appointment  already.

Task-2496629

Part-of: odoo/enterprise#24744