[FIX] sale_subscription: fix journal not being copied from quotation template
Currently when we create a subscription using a quotation template that has a
journal_id defined and the invoices for this subscription are created, the
journal_id of the invoice is the default one. However we want it to be the one
defined in quotation template. This commit overwrites the _prepare_invoice
method to use the quotation template journal_id in the invoice, if it exists.

Part-of: odoo/enterprise#27311