[FIX] web_enterprise: mobile control panel scroll crash
Some part of the code of the mobile control panel
have not been adapted to Owl 2 since the merge of
the latest major version of the framework. A ref
to a node must be used to avoid a crash in the
ControlPanel component.
A test has been added to assure the desired
behavior of the control panel.

closes odoo/enterprise#26567

X-original-commit: 11a7384f3d2379d96240134d66a3099f68e0e8b6
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Luca Vitali <luvi@odoo.com>