[IMP] documents_spreadsheet: add option to clear history
The purpose of this commit is to add a menu item in the file menu
(only visible in debug mode) in order to request a snapshot. Requesting
a snapshot will imply to clear the revisions of the current session, so
the history (undo/redo) will be lost, but loading time will be reduce.
So this button should be used when a spreadsheet takes long time to load.

In an ideal world, we should be able to clearly identify when a session
should be closed (when there is no more active users).

closes odoo/enterprise#26598

Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>