[FIX] appointment: display default select default value
When using the previous feature of the browser, the page displays
and uses the last values of the page. This leads to a missmatch between
what's displayed to the user and used to compute the slots availabilities.

Before the fix:
We open the page to select a slot from the availabilities to book an appointment.
We select a particular employee/user before selecting a slot.
When arriving in the form, the user changes his mind and goes back to the previous page.
On this page, the last employee/user is displayed but the one used to compute the slot
availabilities is in fact the first option of the select.

Now when going back we don't use the previous values of the user and use the default one.

x-original-commit 72d9b8900ffdc9a33ba1ae8109d422bd283a3d8d

task-2812466

X-original-commit: a3dd979ff5f8c818525c3d97edf3eecde7fb56e1
Part-of: odoo/enterprise#27626