[FIX] barcodes_mobile: virtual keyboard could be sticky
It's a non deterministic problem. Sometimes in the barcode application
when scanning, the virtual keyborad pop and never close itself.

It happens because `setTimeout(mobile.methods.closeVirtualKeyboard, 0);`
not always works (I don't know why) and the `_blurBarcodeInput` only
empty the content but doesn't call blur since the keyboard should
already be close due to `closeVirtualKeyboard`.

The current fix, remove the override, in order to still calling
blur if the `closeVirtualKeyBoard` failed. A better fix would be to
ensure that the virtual keyboard on mobile never show

opw-2612019

closes odoo/enterprise#26973

X-original-commit: ba001eadb4d98b9e75c0b26c1199cc28007f5dd1
Signed-off-by: Arnold Moyaux <arm@odoo.com>