[FIX] account_report: keep the context for the DIOT xlsx report
In l10n_mx_report, when printing the xlsx for the DIOT report for a date range,
it will take the lines for the period and all the following lines after the end
of the range. This happens because there is no date_to in the context when reaching
_query_get() in account_move.py

This fix adds the ctx so that it can be used later, especially in the DIOT xlsx.
This appeared in v14 and is present in v15 and saas-15.3
opw-2733346 opw-2709349 opw-2735298

closes odoo/enterprise#27068

X-original-commit: 9f819bfd68bb1f8662cd7eb2f18cf00f3d668c97
Signed-off-by: Quentin De Paoli <qdp@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>