[IMP] documents_spreadsheet: warn cross tab bus synchronisation problem
It can happen (somewhat randomly) that your spreadsheet will indefinitely
display "Saving..." on the top righ corner, in the topbar.

Spreadsheet real time synchronisation relies on the polling cross tab bus.
There's actually only one tab actually polling. It then transmit bus events
to other tabs through local storage.
There's one major caveat: if the master tab is in the background, the browser
can decide to throttle it and set it to "sleep". We don't have any control
over it. If that happens, real time synchronisation on all other tabs is
stopped.

With this commit, we try to mitigate the issue by warning the user that
something is wrong and give him a hint to fix the issue (close other tabs).

This problem should be fixed when we will finally use a websocket bus which
won't have this crosstab bus limitation. The websocket bus has been
under developpent for a while now, but things seems very slow to move and we
don't have any ETA and I'm fed up of this infinite "Saving..." showing up
every so often.

There's also a bug when all tabs are in background and a new election
is triggered: it "forgets" to poll previously polled channels (channels: []).
This should be fixed, we don't have the resources to invest time right now.

closes odoo/enterprise#29527

X-original-commit: 2382a4ba823f7783987a16a289cee5769e197c22
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>