[IMP] approvals: use messaging rpc
This PR prepares the ground for the one introducing the new env in discuss.
The signature of the rpc function has changed a lot from the one in the old env.
In order to ease the transition and reduce the noise, the discuss models now
rely on the messaging's rpc function.

task-2582313

closes odoo/enterprise#26676

Related: odoo/odoo#89858
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>