[FIX] industry_fsm_{sale,stock} : remove product from task
Steps :
Create a FSM task.
Add then remove a product.

Issue :
Error : picking_ids is not a field of sale.order.

Cause :
This field is added to the model with sale_stock, which is not
installed. Therefore, it should be called in industry_fsm_stock i/o
industry_fsm_sale.

opw-2834721

closes odoo/enterprise#27352

X-original-commit: 3f0bd2fd79fa6d0f6c95ead090637e11e70c26ed
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Signed-off-by: Onockx Audric (auon) <auon@odoo.com>