[FIX] account_avatax{,_sale}: calculate tax based on delivery address
Indicated by the 'shipTo' name in the 'addresses' part in an Avatax
transaction API request:

{
  "addresses": {
    ...
    "shipTo": {
      ...
    }
  }
  ...
}

The sale module adds partner_shipping_id on both account.move and
sale.order so always use that if sale is installed.

Using _get_invoice_delivery_partner_id() like in account_taxcloud was
considered but it seems more understandable to directly use the
partner on the record rather than sending a child delivery address to
Avatax without the user knowing this.

opw-2834229

closes odoo/enterprise#26782

X-original-commit: 1135c38c2e998a10a3b3b7db21d93a13b1d6f0e2
Signed-off-by: William André (wan) <wan@odoo.com>