[FIX] account_reports: allow searching account report with hierarchy
A bug was introduced in [commit 1] when using the search bar on certain account
reports. The bug could previously be observed by using the search functionality
on reports based on the account.coa.report model and having the hierarchy
option activated. When this option is activated, searching should be done on
the account groups, not on the individual accounts (account lines are toggled
through their data-parent-id instead). However as of [commit 1] searching is
forced on these lines by including the o_account_searchable_line in the line
dictionary created by the account.coa.report model. When using the search bar
a stacktrace results in the filter_accounts method because the account lines
do not include the account_report_line_name class.

A fix was introduced in [commit 2] that circumvents the error by using an empty
name when the line doesn't include one. The stacktrace is now avoided but the
root problem is still there, resulting in some strange behavior: now, whenever
the end user uses the search bar, all account lines are filtered out.

This commit fixes the search functionality for account reports with the
hierarchy option activated. The account_report_line_name class is now set on
all line types. Additionally parents of searchable and matching child lines
will now always be shown.

[commit 1]: https://github.com/odoo/enterprise/commit/f42450b20d3053177449b196c20d726412768c0a
[commit 2]: https://github.com/odoo/enterprise/commit/55b68e240d495c3db41d47c8d7b384d341af1a86

opw-2823349

X-original-commit: 8002d0b3bb82c45a6f3d4d7585ee520764e645bf
Part-of: odoo/enterprise#27207