[FIX] l10n_mx_reports: refunds column diot report
The diot report was completely wrong, as we did
not have a refunds column, in which should have
be present the amounts of the refunds paid with
following taxes :
IVA 16%, IVA 8%, IVA 0%, Whithheld taxes.

We made a little refactor in _get_lines method
in order to make it a little more readable,
and allow to processcredit notes from above
taxes on last column.
Add also the refund column on diot and dpiva txt
reports.

opw-2733076
opw-2751671

closes odoo/enterprise#29667

X-original-commit: 1e406fcf8755d9abb0c5dfaa3fd4cf1f6ad40ccf
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>