[FIX] project_enterprise: flush before doing SQL query
Purpose
=======
Now that the <mail.mail> are removed in a CRON, one test in project
failed. The reason is that we do not flush the model before doing SQL
queries on them.

Before, after the deletion of the <mail.mail>, everything was flushed
and that's why it worked.

Task-2587345

closes odoo/enterprise#24266

Related: odoo/odoo#73271
Related: odoo/upgrade#3481
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>