[IMP] appointment: auto generate videocall_location
Currently, when a calendar event is generated from an appointment that has no
`location_id` set (which means it is handled online), the calendar event does
not have `videocall_location` set by default, so the salespersons have to
manually send a follow-up to customers on every booked meetings (to share the
URL where online meeting will take place).

This commit improves the behavior for such case, and auto-generates the meeting
URL(videocall_location) for appointments that do not have location set, so that
the URL will be included in the invitation mail being set to the attendees.

Apart from that, this commit also done below changes:

- At the time of booking an appointment, if the appointment has a
  `videocall_location`, it will display the label "How to join"
  with the online meeting URL.

- Change the "Meeting URL" label to "How to Join".

- Change the "https://meet.jit.si/odoo-xyz" label to
  "www.mycompany.com/calendar/join_videocall/xyz".

- Adds the label "Join with Odoo Discuss" if the `base_url` is found in
  `videocall_lcation`, otherwise "Join At" instead of simply showing the
  meeting URL in calendar event related mail templates and moves the
  meeting URL to the bottom of this label.

- To avoid a tiny calendar badge, add style `min-width: 130px`; to the
  calendar badge.

- To make the code more robust and, more importantly, testable,
  replace `request.session ['timezone']` with `request.session.get('timezone')`.

taskID-2821957

closes odoo/enterprise#27136

Related: odoo/odoo#92628
Related: odoo/upgrade#3558
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>