[FIX] l10n_co_edi: remove a tag from html note for no double url
Step to reproduce:
-Install and set up accounting with colombian localization and l10n_co_edi
-Create an invoice with a note that contains a full URL and confirm it
-print and check the xml file generated

Expected behavior:
The  <NOT_1>3.- field contains the url once

Current behavior:
The  <NOT_1>3.- field contains twice the url

Explanation:
The function html2plaintext called on the narration removes the a tag but
not the href field so the URL appears twice. To solve the issue we remove
the <a ...> tags before applying html2plaintext.

opw-2794739

closes odoo/enterprise#27372

X-original-commit: 7b1bf20f3228fd154bea008efebf9604832a9ce7
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Fockedey Martin (mafo) <mafo@odoo.com>