[FIX] account_reports: tax report with misc entry
Goal:
Add a test ensuring the global tax report is working correclty even with
miscellaneous move in which there are an invoice line and a refund line with the
same tax.

reproduction step pre-requisite:
- a simple sale tax A
- miscellaneous contains 2 lines,
	- amount X debit with sale tax A
	- amount Y credit with purchase tax A
- 2 tax lines gets automatically created

Before this commit:
The global tax report will display a base amount of 2*(X-Y) which is wrong.

After this commit:
The global tax report will display a base amount of (X-Y) which is correct.

Context:
The tax details query was giving back twice the amount as result for the base
amount coming from misc move in which you have 2 opposite lines with both having
the same tax.  This is making the generic tax report taking into account twice
the base of the lines.

Solution:
Fixing the mapping in the tax details

closes odoo/enterprise#26961

Community-pr: https://github.com/odoo/odoo/pull/90399
Enterprise-pr: https://github.com/odoo/enterprise/pull/26827
Ticket: 2832745
X-original-commit: 3e398fba350b2190a6306620e85770e804683b10
Related: odoo/odoo#90563
Signed-off-by: Laurent Smet <las@odoo.com>
Signed-off-by: Brice Bartoletti <bib@odoo.com>