[FIX] planning: improve performance of _compute_working_days_count
Before this commit, the shifts are recomputed when a leave is added in
the planned dates of those shifts and the compute of working_days_count
is too costly. This method slows down the validation of a leave or the
creation/suppression of a leave too.

This commit improves the performance of this compute to take less time
as possible to recompute the working_days_count field of those shifts
impacted by a changes on a leave in their planned dates. Also, this
commit checks the `vals` when a leave is updated to only recompute the
`working_days_count` and the `allocated_hours` fields of shifts when the
`resource_id`, `start_datetime` and `end_datetime` fields are changed.

Without the fix, the query counter is equal to 395 queries for
`_recompute_shifts_in_leave_periods` method and with the fix, only 5
queries are counted.

Related PR: #24619

closes odoo/enterprise#26798

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>