[FIX] voip: faster searches on incoming calls
When an incoming call event is triggered, the voip 'user agent'
will contact the backend to search for a user or partner matching
the incomgin call's phone number (formatted in different ways).

A 'limit: 1' was added to (presumably) increase performance, but
it may interact badly with indices set on the phone number fields
to increase performance - and indices on the partner table
in general.

Since the client code is already able to process more than one
result, simply remove the limit.

closes odoo/enterprise#26896

X-original-commit: fa643db32d4461e53cf941812ab905f1daf56f9c
Signed-off-by: Sébastien Theys (seb) <seb@odoo.com>