[FIX] account_reports: allow entries to be foldable in reports (Safari)
When opening the Aged Receivable reports in Safari, the elements in the table
can no longer be foldable once the user started to scroll on the table.

Step to reproduce the issue:
1. Open Safari
2. Install the accounting app
3. Create many (at least 20) invoices on different customers
4. Go to Accounting > Reporting > Age receivable
5. Unfold different customer so that you are forced to scroll in the page (as
there are too many entries to view them all on a single page). If you are lazy,
you can shrink your window size so that you are forced to scroll with less
entries.
You will see that you can no longer unfold other entries.

Solution: In order to have a scrollable (in the Y-axis) table, a trick was
done to clone the original table and keeps only the header. For some reason,
`pointer-events` are working fine in Google Chrome but not in Safari.
Consequently, the `pointer-events` CSS property has been enforced to ensure
reliability (we only allowed `pointer-events` on the header of the table).
Furthermore, another issue was present in Safari. When unfolding/folding, the
text of each row disappeared (only in Safari). Upon investigation, the issue
comes from the `position: relative` which pose issues on Safari. Consequently,
and as the property is not useful in itself (due to parent properties), we
can simply remove it to solve the issue.

opw-2830147

closes odoo/enterprise#27226

X-original-commit: f3a457392a25fba95a7ec1b3da6595869aeee4ba
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Desausoi Laurent (lade) <lade@odoo.com>